Parse `tf2_msgs/TFMessage` from ROS 2 MCAPs as `Transform3D`s (#12224)
Adds a parser for
[TFMessage](https://docs.ros2.org/foxy/api/tf2_msgs/msg/TFMessage.html).
This message type is a bit special because it contains a variable size
list of individual transforms, where each one has its own header
timestamp.

The parser creates `Transform3D`s from that, and assigns each output
transform a `ros2_timestamp` set to the individual transform's header
stamp and a `log/publish_time` timepoint according to the input message.

To achieve this, the default log/pub timepoint handling had to be
slightly adapted to allow a custom override for this parser that can
produce that variable amount of log/pub timepoints.

Closes: https://linear.app/rerun/issue/RR-3181