Remove stablediffusion since source repo has evaporated (#12257)
With the source repo missing and trying to release today my proposal is
to remove the example with a too to revisit. At first I just removed the
links in our docs but then saw we actually reference some images from
the missing repo.

Verified links fixed (for stablediffusion issue) with `pixi run
link-check`

https://github.com/rerun-io/rerun/issues/12255