make fetching chunks in the dfn table provider concurrent while maintaining order (#12253)
### What

Sending 1 request per segment sequentially has a big impact on
performance. The reason we did it this way is ordering, having that
guarantee is beneficial in some common query patterns.

This PR keeps that promise (segment ordering), while performing
concurrent fetches in batches. We batch the requests and execute desired
number of them concurrently and then we re-order responses before
flushing them downstream.

This brings 5-10x (and in some cases more) better benchmark times
(server side change is also required to achieve that), see numbers in
[here](https://linear.app/rerun/issue/RR-3150/benchmarking-and-datset-size-numbers).

We will do a follow up to make target batch size and concurrency level
configurable, for now went with safe and performant defaults.